# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Vinted API",
    "description": "Designed for Le Reacteur Bootcamp",
    "version": "1.0.1"
  },
  "host": "site--backend-vinted--fg6zdpvl2w9z.code.run",
  "basePath": "/",
  "schemes": [
    "HTTPS"
  ],
  "paths": {
...
    "/user/signup": {
      "post": {
        "summary": "Create a new user",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "newsletter": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "Log a user",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}